fix: scope Codex Spark optional extra metric gating
Review fix (#1195): the showOptionalCreditsAndExtraUsage guard was added
to the shared extraRateWindowMetrics(...) helper, which hid every
provider's UsageSnapshot.extraRateWindows when optional extras were off —
regressing Antigravity per-model quotas and Factory core windows.

Gate only the Codex provider's extra windows (Codex Spark and any future
Codex additional limit are optional extra usage); other providers' extra
windows always render. Codex primary/session and weekly lanes are
unaffected, and percent/reset behavior is preserved.

Also fixes CI lint-build-test: split the Codex additional_rate_limits /
Spark parser tests out of CodexOAuthTests into CodexAdditionalRateLimitsTests
to clear a SwiftLint type_body_length violation (struct body was 881 > 800).

Adds an Antigravity regression test asserting per-model extra windows
still render when showOptionalCreditsAndExtraUsage == false.

Refs #1177
